// apps/frontend/src/api/generated/model/updateSendMessageRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Alittlebyte MVP API
 * Versioned JSON contract consumed by the Angular SPA.
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateSendMessageRequest {
  /**
     * @maxLength 255
     * @nullable
     */
  recipient?: string | null;
  /**
     * @maxLength 255
     * @nullable
     */
  subject?: string | null;
  /**
     * @maxLength 5000
     * @nullable
     */
  body?: string | null;
}